requirements.txt
aerich==0.7.1
aiocron==1.8
aiofiles==22.1.0
aiosqlite==0.17.0
anyio==3.6.1
asgi-lifespan==2.0.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.9.24
click==8.1.3
croniter==1.3.8
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.85.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
mypy==0.982
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.2
pyinstrument==4.4.0
PyJWT==1.7.1
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.5
tortoise-orm==0.19.2
trio==0.22.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
uvicorn==0.18.3
wrapt==1.14.1
